fix msvc errors and warnings

+++ include/boost/stringify/v0/input_types/int.hpp
@@ -451,7 +451,12 @@ private:
         }
         else
         {
-            write_digits_bigsep(dig_it, grp_buff, grp_it, sep.ch, sep.size);
+            write_digits_bigsep
+                ( dig_it
+                , grp_buff
+                , grp_it
+                , sep.ch
+                , static_cast<unsigned>(sep.size) );
         }
     }
